add pipeline class so all history can be retained
